Ajout du système de collision

+++ source/src/main/java/com/mauja/maujaadventures/modele/Position.java
@@ -1,15 +1,15 @@
 package com.mauja.maujaadventures.modele;
 
 public class Position {
-    private int positionX;
-    private int positionY;
+    private double positionX;
+    private double positionY;
     /**
      * Constructeur de la classe Position
      * @param positionX Position de l'objet en X
      * @param positionY Position de l'obejt en Y
      * @author Tremblay Jeremy, Vignon Ugo, Viton Antoine, Wissocq Maxime, Coudour Adrien
      */
-    public Position(int positionX, int positionY) {
+    public Position(double positionX, double positionY) {
         this.positionX = positionX;
         this.positionY = positionY;
     }
@@ -18,23 +18,23 @@ public class Position {
      * @return La position X de l'objet
      * @author Tremblay Jeremy, Vignon Ugo, Viton Antoine, Wissocq Maxime, Coudour Adrien
      */
-    public int getPositionX() { return positionX; }
+    public double getPositionX() { return positionX; }
     /**
      * Setter de la position X
      * @param positionX Nouvelle position X de l'objet
      * @author Tremblay Jeremy, Vignon Ugo, Viton Antoine, Wissocq Maxime, Coudour Adrien
      */
-    public void setPositionX(int positionX) { this.positionX = positionX; }
+    public void setPositionX(double positionX) { this.positionX = positionX; }
     /**
      * Getter de position Y
      * @return Position Y de l'objet
      * @author Tremblay Jeremy, Vignon Ugo, Viton Antoine, Wissocq Maxime, Coudour Adrien
      */
-    public int getPositionY() { return positionY; }
+    public double getPositionY() { return positionY; }
     /**
      * Setter de la position Y
      * @param positionY Nouvelle position Y de l'objet
      * @author Tremblay Jeremy, Vignon Ugo, Viton Antoine, Wissocq Maxime, Coudour Adrien
      */
-    public void setPositionY(int positionY) { this.positionY = positionY; }
+    public void setPositionY(double positionY) { this.positionY = positionY; }
 }
